split contetx.py in download_ctx.py and input_ctx.py; split strategy.py in idownload.py and iinput.py; remove parameter definition in download_ctx.py, now only download declare them and daccess can overwrite them
