--- download/context.py
@@ -1,78 +1,0 @@
-from download.strategy import DownloadStrategy, InputStrategy
-
-
-class DownloadContext:
-    """
-        The Context defines the interface of interest to clients.
-        """
-
-    def __init__(self, strategy: DownloadStrategy) -> None:
-        """
-        Usually, the Context accepts a strategy through the constructor, but
-        also provides a setter to change it at runtime.
-        """
-
-        self._strategy = strategy
-
-    @property
-    def strategy(self) -> DownloadStrategy:
-        """
-        The Context maintains a reference to one of the Strategy objects. The
-        Context does not know the concrete class of a strategy. It should work
-        with all strategies via the Strategy interface.
-        """
-
-        return self._strategy
-
-    @strategy.setter
-    def strategy(self, strategy: DownloadStrategy) -> None:
-        """
-        Usually, the Context allows replacing a Strategy object at runtime.
-        """
-
-        self._strategy = strategy
-
-    def download(self, dataset, working_domain, fields, return_type, in_memory=False, rm_file=False, max_attempt=5):
-        return self._strategy.download(dataset, working_domain, fields, in_memory, rm_file, max_attempt=max_attempt,
-                                       return_type=return_type)
-
-
-class InputContext:
-    """
-        The Context defines the interface of interest to clients.
-        """
-
-    def __init__(self, strategy: InputStrategy) -> None:
-        """
-        Usually, the Context accepts a strategy through the constructor, but
-        also provides a setter to change it at runtime.
-        """
-
-        self._strategy = strategy
-
-    @property
-    def strategy(self) -> InputStrategy:
-        """
-        The Context maintains a reference to one of the Strategy objects. The
-        Context does not know the concrete class of a strategy. It should work
-        with all strategies via the Strategy interface.
-        """
-
-        return self._strategy
-
-    @strategy.setter
-    def strategy(self, strategy: InputStrategy) -> None:
-        """
-        Usually, the Context allows replacing a Strategy object at runtime.
-        """
-
-        self._strategy = strategy
-
-    def get_wd(self, workingDomain, dataset) -> dict:
-        """
-        @param workingDomain: dict with spatial/time information:
-                lonLat: list of list, the internal list has the format:  [[minLon , maxLon], [minLat , maxLat]]
-                depth: depth range in string format: [minDepth, maxDepth]
-                time: list of two strings that represent a time range: [YYYY-MM-DDThh:mm:ssZ, YYYY-MM-DDThh:mm:ssZ]
-        """
-        return self._strategy.get_wd(workingDomain, dataset)
